// src/app/administration/administrative-report/administrative-report.interface.ts
export interface AdministrativeReport {
    no: string;
    date: string;
    area: string;
    locality: string;
    detail: string;
    description: string;
    latitude: string;
    longitude: string;
    observations: string;
    recommendationsNavigator: string;
    source: string;
    preparedBy: string;
}